Reduce size of UniRef data in database
Fully parsed Uniref90 dataset consists of over 300M records and
that data is killing any queries to the UniprotIndex table. This
modification is a first step in reducing the number of records.
It accomplishes this with code that:

1. Queries the sqlite_db for actual UniRef hit accessions
2. Identifies all records in the fully parsed UniRef .tsv that
have accessions in the database
3. Saves those records to a new "reduced" file

Also a new Sqlite_connect subroutine was generated to return an
array from a single SQLite column.